+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/Cascading.java
@@ -17,5 +17,13 @@
 package io.fabric8.kubernetes.client.dsl;
 
 public interface Cascading<T> {
+  /**
+   * deletes dependent resources. Sets `orphanDependents` field to `false` when set `true`
+   *
+   * @deprecated Please Use {@link io.fabric8.kubernetes.client.PropagationPolicyConfigurable} instead. This field has been deprecated since 1.7.
+   * @param enabled whether dependents should be orphaned or not.
+   * @return resource
+   */
+  @Deprecated
   T cascading(boolean enabled);
 }