+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/internal/ClusterRoleOperationsImpl.java
@@ -32,7 +32,7 @@ public class ClusterRoleOperationsImpl extends HasMetadataOperation<ClusterRole,
   }
 
   public ClusterRoleOperationsImpl(OkHttpClient client, Config config, String namespace) {
-    this(new OperationContext().withOkhttpClient(client).withConfig(config).withNamespace(namespace));
+    this(new OperationContext().withOkhttpClient(client).withConfig(config).withNamespace(namespace).withPropagationPolicy(DEFAULT_PROPAGATION_POLICY));
   }
 
   public ClusterRoleOperationsImpl(OperationContext context) {